// api/model/findFeedCosmeticsDto.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * POCHY API 명세서
 * POCHY Swagger UI
 * OpenAPI spec version: 1.0.0
 */

export interface FindFeedCosmeticsDto {
  cosmeticId?: number;
  productImageUrl?: string;
  name?: string;
  brand?: string;
  memo?: string;
  category?: string;
  subCategory?: string;
  url?: string;
}